Split main to reduce per file loc limit
